Move public API Header to InfluxDB/ subdirectory
