+++ script/include_library/main.cxx
@@ -1,5 +1,5 @@
 // Sample project to check if the deployment process works
-#include <InfluxDBFactory.h>
+#include <InfluxDB/InfluxDBFactory.h>
 
 int main()
 {
