+++ src/UnixSocket.cxx
@@ -26,7 +26,7 @@
 ///
 
 #include "UnixSocket.h"
-#include "InfluxDBException.h"
+#include "InfluxDB/InfluxDBException.h"
 #include <string>
 
 namespace influxdb::transports
