+++ test/HttpTest.cxx
@@ -21,7 +21,7 @@
 // SOFTWARE.
 
 #include "HTTP.h"
-#include "InfluxDBException.h"
+#include "InfluxDB/InfluxDBException.h"
 #include "mock/CprMock.h"
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/trompeloeil.hpp>
