+++ test/ProxyTest.cxx
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include "Proxy.h"
+#include "InfluxDB/Proxy.h"
 #include <catch2/matchers/catch_matchers_all.hpp>
 #include <catch2/catch_test_macros.hpp>
