+++ test/system/SystemTest.h
@@ -22,8 +22,8 @@
 
 #pragma once
 
-#include "InfluxDBFactory.h"
-#include "InfluxDBException.h"
+#include "InfluxDB/InfluxDBFactory.h"
+#include "InfluxDB/InfluxDBException.h"
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_all.hpp>
 #include <string>
